+++ Sources/MapboxCoreNavigation/EHorizon/RoadGraph.swift
@@ -6,6 +6,8 @@ import MapboxNavigationNative
  `RoadGraph` provides methods to get edge shape (e.g. `RoadGraph.Edge`) and metadata.
  
  You do not create a `RoadGraph` object manually. Instead, observe the `Notification.Name.electronicHorizonDidUpdatePosition` notification to obtain edge identifiers and get more details about the edges using the `RouteController.roadGraph` or `PassiveLocationManager.roadGraph` property.
+ 
+ Note: The Mapbox Electronic Horizon feature of the Mapbox Navigation SDK is in public beta and is subject to changes, including its pricing. Use of the feature is subject to the beta product restrictions in the Mapbox Terms of Service. Mapbox reserves the right to eliminate any free tier or free evaluation offers at any time and require customers to place an order to purchase the Mapbox Electronic Horizon feature, regardless of the level of use of the feature.
  */
 public final class RoadGraph {