+++ Sources/MapboxCoreNavigation/EHorizon/RoadGraphPosition.swift
@@ -3,7 +3,10 @@ import MapboxNavigationNative
 
 extension RoadGraph {
 
-    /** The position of a point object in the road graph. */
+    /** The position of a point object in the road graph.
+     
+        Note: The Mapbox Electronic Horizon feature of the Mapbox Navigation SDK is in public beta and is subject to changes, including its pricing. Use of the feature is subject to the beta product restrictions in the Mapbox Terms of Service. Mapbox reserves the right to eliminate any free tier or free evaluation offers at any time and require customers to place an order to purchase the Mapbox Electronic Horizon feature, regardless of the level of use of the feature.
+     */
     public struct Position {
 
         /** The edge identifier along which the point object lies. */