+++ Sources/MapboxCoreNavigation/EHorizon/RoadObjectEdgeLocation.swift
@@ -7,6 +7,8 @@ extension RoadObject {
      Represents location of road object on road graph.
      
      A point object is represented by a single edge whose location has the same `fractionFromStart` and `fractionToEnd`.
+     
+     Note: The Mapbox Electronic Horizon feature of the Mapbox Navigation SDK is in public beta and is subject to changes, including its pricing. Use of the feature is subject to the beta product restrictions in the Mapbox Terms of Service. Mapbox reserves the right to eliminate any free tier or free evaluation offers at any time and require customers to place an order to purchase the Mapbox Electronic Horizon feature, regardless of the level of use of the feature.
      */
     public struct EdgeLocation {